fix: 🐛 fix yarn build module
