+++ src/app/network-dao/proposal-list/_page.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import DaoDetail from '../../dao/[daoId]/_page'
+
+export default function DeoDetails() {
+  return (
+    <div>
+    <DaoDetail />
+  </div>)
+}
